hiera: set python::provider to pip
The yum package 'python-pip' was renamed to 'python2-pip'; as a
result, `yum list` can no longer find the package and therefore
puppet refuse to install it.

Setting provider to pip will allow pip to be self-managed. Although
the bootstraping process is slightly less "trustworthy", it is yet
another domain we have to rely on.